update text about PyWorking, add junior.guru
We talked about this with @vahalova and the rest of the PyLadies Brno crew at Pyvo in Brno in October 2019. Then I checked with others on Slack - https://pyvec.slack.com/archives/C1MJT7TGT/p1576013513189600 The response was positive. It took me just a few months, but here I am, finally sending a proposal how it could look like.